+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- (FSEUS-368) Add Custom NSU to Payer Auth Response
+
 ## [1.25.0] - 2024-07-30
 
 ### Changed